% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2gene.R
\name{calc_binary_peak2gene}
\alias{calc_binary_peak2gene}
\title{Calculate Binary Peak-to-Gene Matrix}
\usage{
calc_binary_peak2gene(peak2gene_distance, distance_thr)
}
\arguments{
\item{peak2gene_distance}{A sparse matrix where non-zero entries represent distances between peaks and genes.}

\item{distance_thr}{Numeric value specifying the maximum distance threshold for peak-to-gene associations.}
}
\value{
A binary sparse matrix of the same dimensions as `peak2gene_distance` with entries of 1 where
        peak-to-gene distances are within `distance_thr` and 0 elsewhere.
}
\description{
This function generates a binary sparse matrix indicating peak-to-gene associations
within a specified distance threshold.
}
