% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_buster_score.R
\name{calc_pv_cbscore}
\alias{calc_pv_cbscore}
\title{Transform Cluster-Buster Scores to P-values}
\usage{
calc_pv_cbscore(cbscore, mapping_mat, dir_null_cbscore, n_control, ncores)
}
\arguments{
\item{cbscore}{Matrix of Cluster-Buster scores with motifs as rows and consensus peaks as columns.}

\item{mapping_mat}{Sparse matrix mapping consensus peaks to control regions. See \code{\link{extract_matched_control_regions}}.}

\item{dir_null_cbscore}{Directory path with pre-computed null distributions for each motif.}

\item{n_control}{Number of control regions used in null score calculation in alignment with mapping_mat.}

\item{ncores}{Number of cores for parallel processing.}
}
\value{
A P-value matrix.
}
\description{
Transforms Cluster-Buster scores into P-values by comparing to (motif, consensus peak)-specific null scores.
}
