% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF2gene.R
\name{calc_tf2gene_score}
\alias{calc_tf2gene_score}
\title{Calculate TF-Gene Score}
\usage{
calc_tf2gene_score(tf2peak, peak2gene_distance, distance_thr)
}
\arguments{
\item{tf2peak}{A matrix of TF-to-peak scores.}

\item{peak2gene_distance}{A sparse matrix with peak-to-gene distances.}

\item{distance_thr}{Numeric. The maximum peak-to-gene distance threshold.}
}
\value{
A matrix representing TF-to-gene scores.
}
\description{
Computes the transcription factor (TF)-to-gene score by combining TF-peak
scores with peak-to-gene distances, using a specified distance threshold.
}
