% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF2peak.R
\name{calc_tf2peak_score}
\alias{calc_tf2peak_score}
\title{Calculate TF-Peak Score}
\usage{
calc_tf2peak_score(
  trans_cbscore,
  n_control,
  mat_tf2motif = mat_tf2motif,
  tf2motif_level
)
}
\arguments{
\item{trans_cbscore}{A matrix of transformed Cluster-Buster scores, i.e., P-values, of enriched motifs.
See \code{\link{transform_cbscore}} and \code{\link{select_active_motifs_with_cbscore}}.}

\item{n_control}{Number of control regions used to generate \code{trans_cbscore}.}

\item{mat_tf2motif}{A binary matrix indicating TF-motif connections. See \code{?mat_tf2motif} for default.}

\item{tf2motif_level}{Integer (≤3) indicating the minimum level of TF-motif connections to retain.
1: by similarity; 2: by orthology; 3. direct annotation.}
}
\value{
A matrix representing TF-to-peak scores.
}
\description{
Computes a transcription factor (TF)-to-peak score with transformed Cluster-Buster score matrix
of enriched motifs and TF-motif connection matrix.
}
