% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_active_motifs.R
\name{cauchy_combination_test}
\alias{cauchy_combination_test}
\title{Cauchy Combination Test}
\usage{
cauchy_combination_test(pvals)
}
\arguments{
\item{pvals}{Numeric vector of p-values (each between 1e-16 and 1).}
}
\value{
Combined p-value from the Cauchy test.
}
\description{
This function combines \( p \)-values using the Cauchy Combination Test, which is effective
for handling dependencies among tests.
}
