% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_buster_score.R
\name{collect_motif_selection}
\alias{collect_motif_selection}
\title{Collect Motif Selection Results}
\usage{
collect_motif_selection(dir_out, ncores)
}
\arguments{
\item{dir_out}{A character string specifying the directory containing the RDS files.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}
}
\value{
A tibble with columns motif, logfc, and pv.
}
\description{
This function collects outputs of \link{\code{select_motifs_with_cbscore}} from a specified directory.
}
