% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_tf2motif}
\alias{df_tf2motif}
\title{TF-motif annotations from cisTarget}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 253096 rows and 14 columns.
}
\usage{
df_tf2motif
}
\description{
A tibble annotating connections between TF and motifs.
}
\keyword{datasets}
