% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matched_control_regions.R
\name{extract_best_matched_control_region}
\alias{extract_best_matched_control_region}
\title{Extract Best Matched Control Regions}
\usage{
extract_best_matched_control_region(mapping_mat, ncores)
}
\arguments{
\item{mapping_mat}{A sparse matrix where rows represent consensus peaks and columns
represent control regions. Each entry in the matrix is a value indicating the
distance between the consensus peak and the control region.
See \code{\link{extract_matched_control_regions}}.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}
}
\value{
A named vector of the best control region for each consensus peak.
}
\description{
This function finds the best-matched control region for each consensus peak.
}
