% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matched_control_regions.R
\name{extract_top_matched_control_regions}
\alias{extract_top_matched_control_regions}
\title{Extract Top Matched Control Regions}
\usage{
extract_top_matched_control_regions(mapping_mat, n_control, ncores)
}
\arguments{
\item{mapping_mat}{A sparse matrix where rows represent consensus peaks and columns
represent control regions. Each entry in the matrix is a value indicating the
distance between the consensus peak and the control region.
See \code{\link{extract_matched_control_regions}}.}

\item{n_control}{Number of matched control regions to extract for each consensus peak.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}
}
\value{
A named list of the top matched control regions for each consensus peak.
}
\description{
This function finds the top-matched control regions for each consensus peak.
}
