% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{format_consensus_peaks}
\alias{format_consensus_peaks}
\title{Format ATAC-seq consensus peak}
\usage{
format_consensus_peaks(fbed)
}
\arguments{
\item{fbed}{A character string specifying the path to the BED file.}
}
\value{
A tibble with four columns:
\describe{
  \item{chr}{Chromosome identifier.}
  \item{start}{Start position of the peak.}
  \item{end}{End position of the peak.}
  \item{openness}{Openness score of the peak.}
}
}
\description{
This function reads a BED file and formats the consensus peaks. It only retains peaks in chromosome 1-22, X, & Y.
}
