% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomputed-control_regions.R
\name{generate_control_regions}
\alias{generate_control_regions}
\title{Generate Genome-Wide Control Regions}
\usage{
generate_control_regions(
  meta_gene = TF2gene::grch38,
  chr_lens = TF2gene::grch38_chr_lens,
  region_length = 500,
  radius = 5e+05
)
}
\arguments{
\item{meta_gene}{A tibble containing gene metadata with columns: 'id', 'symbol', 'biotype', 'chr', 'start', 'end', 'strand', 'tss' (default is TF2gene::grch38).}

\item{chr_lens}{A tibble containing chromosome lengths with columns: 'chr' and 'length' (default is TF2gene::grch38_chr_lens).}

\item{region_length}{A numeric value specifying the length of each control_region (default is 500).}

\item{radius}{A numeric value specifying the radius around the TSS to consider for generating control regions (default is 500,000).}
}
\value{
A tibble with columns: 'chr', 'start', and 'end' representing the generated control regions.
}
\description{
This function generates control regions encompassing a given set of genes. It creates intervals around gene transcription start sites (TSS), resizes them, and splits them into non-overlapping windows of a specified length.
}
