% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{get_cb_motif_info}
\alias{get_cb_motif_info}
\title{Get Cluster-Buster Motif Information}
\usage{
get_cb_motif_info(dir_motif, ncores, verbose = TRUE)
}
\arguments{
\item{dir_motif}{A character string specifying the directory containing motif files.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{verbose}{A logical value indicating whether to display progress. Default is TRUE.}
}
\value{
A tibble with the following columns:
\describe{
  \item{motif}{A character vector containing motif IDs.}
  \item{num_inner_motif}{An integer vector containing the number of inner motifs for each motif ID.}
}
}
\description{
This function retrieves information about motifs from a specified directory.
}
