% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_peaks.R
\name{get_control_peaks}
\alias{get_control_peaks}
\title{Generate Null Peaks}
\usage{
get_control_peaks(
  meta_gene = grch38,
  chr_lens = hg38_chr_lens,
  peak_length = 500,
  radius = 5e+05
)
}
\arguments{
\item{meta_gene}{A tibble containing gene metadata with columns: 'id', 'symbol', 'biotype', 'chr', 'start', 'end', 'strand', 'tss' (default is grch38).}

\item{chr_lens}{A tibble containing chromosome lengths with columns: 'chr' and 'length' (default is hg38_chr_lens).}

\item{peak_length}{A numeric value specifying the length of each peak (default is 500).}

\item{radius}{A numeric value specifying the radius around the TSS to consider for generating peaks (default is 500,000).}
}
\value{
A tibble with columns: 'chr', 'start', and 'end' representing the generated null peaks.
}
\description{
This function generates control peaks encompassing a given set of genes. It creates intervals around gene transcription start sites (TSS), resizes them, and splits them into non-overlapping windows of a specified length.
}
