% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2gene.R
\name{get_peak2gene}
\alias{get_peak2gene}
\title{Generate Binary Peak-to-Gene Matrix by Distance}
\usage{
get_peak2gene(bed_consensus_peaks, distance_thr = 1e+05, ncores)
}
\arguments{
\item{bed_consensus_peaks}{Path to the BED file containing consensus peaks.}

\item{distance_thr}{Numeric value specifying the maximum distance threshold for peak-to-gene associations. Default is 100 kb.}

\item{ncores}{Number of cores to use for parallel processing.}
}
\value{
A binary sparse matrix indicating peak-to-gene associations.
}
\description{
This function reads consensus peaks from a BED file, calculates the distances between peaks and genes,
and generates a binary sparse matrix indicating peak-to-gene associations within a specified distance threshold.
}
