% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch38}
\alias{grch38}
\title{GRCh38 gene meta data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220435 rows and 8 columns.
}
\usage{
grch38
}
\description{
Genes on chromosome 1-22, X, Y. Extracted from biomaRt on Sep 3, 2024. See data.R for details.
}
\keyword{datasets}
