% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_chr_lens}
\alias{hg38_chr_lens}
\title{Chromosome length of hg38 build}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 2 columns.
}
\usage{
hg38_chr_lens
}
\description{
Only retain chromosome 1-22, X, Y. Documented in data-raw/.
}
\keyword{datasets}
