% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_buster_score.R
\name{load_cbscore}
\alias{load_cbscore}
\title{Load Cluster-Buster Scores}
\usage{
load_cbscore(feather)
}
\arguments{
\item{feather}{A character string representing the path to the feather file.}
}
\value{
A matrix of scores with motifs as rows and regions as columns.
}
\description{
This function loads the cluster-buster scores from a feather file.
}
