% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_buster_score.R
\name{load_null_cbscore}
\alias{load_null_cbscore}
\title{Load Null Cluster-Buster Scores}
\usage{
load_null_cbscore(motif, dir_null_cbscore)
}
\arguments{
\item{motif}{A character string representing the motif name.}

\item{dir_null_cbscore}{A character string representing the directory containing the null cluster-buster scores.}
}
\value{
A named numeric vector of scores.
}
\description{
This function loads the null cluster-buster scores for a given motif.
}
