% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_tf2motif}
\alias{mat_tf2motif}
\title{Processed TF-motif annotations}
\format{
An object of class \code{dgCMatrix} with 1605 rows and 27072 columns.
}
\usage{
mat_tf2motif
}
\description{
A sparse TF by motif matrix. 3: Direct; 2: By orthology; 1: By similarity.
}
\keyword{datasets}
