% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_control_region}
\alias{meta_control_region}
\title{Meta data of genome-wide control regions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5807331 rows and 8 columns.
}
\usage{
meta_control_region
}
\description{
A tibble with columns chr, start, end, dist (distance to nearset TSS), freq_A, freq_C, freq_G, freq_T.
}
\keyword{datasets}
