% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_motif}
\alias{meta_motif}
\title{Metadata of motifs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10249 rows and 2 columns.
}
\usage{
meta_motif
}
\description{
A tibble of the metadata of motifs.
}
\keyword{datasets}
