% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{plot_meta_peak}
\alias{plot_meta_peak}
\title{Plot Metadata of Peaks}
\usage{
plot_meta_peak(meta_peak, facet_by = NULL, add_cor = TRUE)
}
\arguments{
\item{meta_peak}{A tibble containing metadata of peaks, including distance to the nearest TSS and GC content. See \code{\link{summarize_consensus_peaks}}.}

\item{facet_by}{A character string specifying the column to facet by (default is NULL).}

\item{add_cor}{A logical value indicating whether to add correlation test results to the plot (default is TRUE).}
}
\value{
A ggplot object representing the hexbin plot.
}
\description{
This function creates a hexbin plot of the distance to the nearest transcription start site (TSS) versus the GC content of the peak sequences.
}
