% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_buster_score.R
\name{read_cbscore}
\alias{read_cbscore}
\title{Read Cluster-Buster Scores}
\usage{
read_cbscore(feather)
}
\arguments{
\item{feather}{A character string representing the path to the feather file.}
}
\value{
A matrix of scores with motifs as rows and regions as columns.
}
\description{
This function reads the cluster-buster scores from a feather file.
}
