% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrow_peaks.R
\name{read_narrow_peaks}
\alias{read_narrow_peaks}
\title{Read Narrow Peaks File}
\usage{
read_narrow_peaks(file)
}
\arguments{
\item{file}{A string specifying the path to the narrow peaks file.}
}
\value{
A tibble containing the narrow peaks data.
}
\description{
This function reads a narrow peaks file into a tibble.
}
