% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_peaks.R
\name{sample_matched_control_peaks}
\alias{sample_matched_control_peaks}
\title{Sample Matched Control Peaks}
\usage{
sample_matched_control_peaks(
  meta_peak,
  meta_control_peak,
  nbins,
  n_control,
  ncores,
  seed
)
}
\arguments{
\item{meta_peak}{A tibble containing metadata for consensus peaks. See \code{\link{summarize_consensus_peaks}}.}

\item{meta_control_peak}{A tibble containing metadata for control peaks.}

\item{nbins}{Number of bins in each dimension. May be a scalar or a 2 element vector.}

\item{n_control}{A numeric value specifying the number of control peaks to sample.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{seed}{A numeric value specifying the seed for random sampling.}
}
\value{
A tibble of subset of meta_control_peak containing the sampled control peaks.
}
\description{
This function samples control peaks that match the distribution of consensus peaks based on distance to nearest TSS and GC content.
}
