% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_matched_control_regions.R
\name{sample_matched_control_regions}
\alias{sample_matched_control_regions}
\title{Sample Matched control regions}
\usage{
sample_matched_control_regions(
  meta_peak,
  meta_control_region,
  nbins,
  n_control,
  ncores,
  seed
)
}
\arguments{
\item{meta_peak}{A tibble containing metadata for consensus peaks. See \code{\link{summarize_consensus_peaks}}.}

\item{meta_control_region}{A tibble containing metadata for control regions.}

\item{nbins}{Number of bins in each dimension. May be a scalar or a 2 element vector.}

\item{n_control}{A numeric value specifying the number of control regions to sample.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{seed}{A numeric value specifying the seed for random sampling.}
}
\value{
A tibble of subset of meta_control_region containing the sampled control regions.
}
\description{
This function samples control regions that match the A/C/G/T frequencies of consensus peaks.
}
