% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_active_motifs.R
\name{select_active_motifs}
\alias{select_active_motifs}
\title{Select Active Motifs}
\usage{
select_active_motifs(trans_cbscore)
}
\arguments{
\item{trans_cbscore}{Matrix or data frame of transformed Cluster-Buster scores (See \code{\link{transform_cbscore}}).}
}
\value{
Tibble with `motif` and `prop_enrich` columns.
}
\description{
Calculates enrichment proportions for each motif based on transformed Cluster-Buster scores.
}
