% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_active_motifs.R
\name{select_active_motifs_with_cbscore}
\alias{select_active_motifs_with_cbscore}
\title{Select Active Motifs Based on Cluster-Buster Scores}
\usage{
select_active_motifs_with_cbscore(
  cbscore,
  control_regions,
  dir_null_cbscore,
  ncores,
  dir_out
)
}
\arguments{
\item{cbscore}{A matrix of cluster-buster scores of consensus peaks.}

\item{control_regions}{A names vector of best control regions for each consensus peak.
See \code{\link{extract_best_matched_control_region}}.}

\item{dir_null_cbscore}{A character string representing the directory containing the cluster-buster scores on control regions.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{dir_out}{A character string representing the directory to save the results.}
}
\value{
This function is called by its side effect.
}
\description{
This function selects active motifs by comparing their cluster-buster scores on consensus peaks and matched control regions.
}
