% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{select_motifs}
\alias{select_motifs}
\title{Select Motifs Based on P-value and Q-value Rules}
\usage{
select_motifs(pv_cbscore, num_control_per_peak, rules, ncores)
}
\arguments{
\item{pv_cbscore}{A matrix of P-values for motifs to ATAC-seq consensus peaks.}

\item{num_control_per_peak}{Number of control regions per peak.}

\item{rules}{A vector of rules for selecting motifs.}

\item{ncores}{Number of cores to use for parallel processing.}
}
\value{
A vector of selected motif names.
}
\description{
This function selects motifs based on empirical P-value and Q-value rules.
}
