% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_buster_score.R
\name{select_motifs_with_cbscore}
\alias{select_motifs_with_cbscore}
\title{Select Motifs Based on Cluster-Buster Scores}
\usage{
select_motifs_with_cbscore(
  cbscore,
  control_peaks,
  dir_null_cbscore,
  ncores,
  dir_out
)
}
\arguments{
\item{cbscore}{A matrix of cluster-buster scores.}

\item{control_peaks}{A character vector of control peak names.}

\item{dir_null_cbscore}{A character string representing the directory containing the null cluster-buster scores.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{dir_out}{A character string representing the directory to save the results.}
}
\value{
This function is called by its side effect.
}
\description{
This function selects motifs based on their cluster-buster scores and control peaks.
}
