% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{summarize_consensus_peaks}
\alias{summarize_consensus_peaks}
\title{Summarize ATAC-seq Consensus Peaks}
\usage{
summarize_consensus_peaks(fasta, ncores, meta_gene = grch38, verbose = TRUE)
}
\arguments{
\item{fasta}{A character string specifying the path to the FASTA file containing sequences of consensus peaks.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{meta_gene}{A data frame containing metadata of genes, including chromosome and TSS information (default is grch38).}

\item{verbose}{A logical value indicating whether to print progress messages (default is TRUE).}
}
\value{
A tibble with the following columns:
\describe{
  \item{chr}{Chromosome identifier.}
  \item{start}{Start position of the peak.}
  \item{end}{End position of the peak.}
  \item{name}{Name of the peak.}
  \item{center}{Center position of the peak.}
  \item{dist}{Distance from the center of the peak to the nearest gene TSS.}
  \item{gc_content}{GC content of the peak sequence.}
}
}
\description{
This function reads a FASTA file containing sequences of consensus peaks and a metadata table of gene information. It calculates the distance between the centers of the peaks and the nearest gene transcription start site (TSS), and computes the GC content of each peak.
}
