% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_features_of_regions.R
\name{summarize_features_of_regions}
\alias{summarize_features_of_regions}
\title{Summarize Features of Genomic Regions}
\usage{
summarize_features_of_regions(fasta, ncores, meta_gene = grch38)
}
\arguments{
\item{fasta}{A character string specifying the path to the FASTA file containing sequences of consensus peaks.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{meta_gene}{A data frame containing metadata of genes, including chromosome and TSS information (default is grch38).}
}
\value{
A tibble with the following columns:
\describe{
  \item{chr}{Chromosome identifier.}
  \item{start}{Region start position (0-based, included).}
  \item{end}{Region end position (0-based, excluded).}
  \item{dist}{Distance from region center to nearest TSS.}
  \item{freq_A}{Frequency of A.}
  \item{freq_C}{Frequency of C.}
  \item{freq_G}{Frequency of G.}
  \item{freq_T}{Frequency of T.}
}
}
\description{
This function reads a FASTA file containing sequences of genomic regions and a metadata table of gene information.
It calculates the distance between the centers of the regions and the nearest TSS, and the A/C/G/T frequencies of each region.
}
