% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF2gene.R
\name{TF2gene}
\alias{TF2gene}
\title{Calculate TF-Gene P-value Matrix}
\usage{
TF2gene(
  feather_file,
  fasta_consenesus_peaks,
  dir_null_cbscore,
  peak2gene,
  dir_save,
  resume,
  ncores,
  num_control_per_peak = 1000,
  rules = c(0.05, 0.01, 0.25)
)
}
\arguments{
\item{feather_file}{Path to the feather file containing Cluster-Buster scores.}

\item{fasta_consenesus_peaks}{Path to the FASTA file of ATAC-seq consensus peaks.}

\item{dir_null_cbscore}{Directory containing null Cluster-Buster scores.}

\item{peak2gene}{A matrix indicating the connection between ATAC-seq consensus peaks and genes.}

\item{dir_save}{Directory to save intermediate outputs, which will be loaded when the task is resumed.}

\item{resume}{Logical indicating whether to resume the unfinished task.}

\item{ncores}{Number of cores for parallel computation.}

\item{num_control_per_peak}{Number of matched control regions per peak. Default is 1000.}

\item{rules}{A vector of rules for selecting motifs. Default is `c(0.05, 0.01, 0.25)`.}
}
\value{
A list containing (1) a P-value matrix for all TFs to all genes, where 2 indicates TF has no motifs or gene has no peaks,
(2) a compact P-value matrix removing the TFs and genes with P-values of 2, (3) a compact Q-value matrix where Q-values are calculated for each TF.
}
\description{
This function calculates the P-value matrix for transcription factors (TFs) and genes indicating the regulatory relationship.
}
