% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_cluster_buster_score.R
\name{transform_cbscore}
\alias{transform_cbscore}
\title{Transform Cluster-Buster Scores}
\usage{
transform_cbscore(
  cbscore,
  depleted_motifs,
  enriched_motifs,
  mapping_mat,
  dir_null_cbscore,
  n_control,
  ncores
)
}
\arguments{
\item{cbscore}{Matrix of Cluster-Buster scores with motifs as rows and consensus peaks as columns.}

\item{depleted_motifs}{Vector of motif names set as depleted, with scores set to zero.}

\item{enriched_motifs}{Vector of motif names to be enriched and transformed.}

\item{mapping_mat}{Sparse matrix mapping consensus peaks to control regions. See \code{\link{extract_matched_control_regions}}.}

\item{dir_null_cbscore}{Directory path with precomputed null distributions for each motif.}

\item{n_control}{Number of control regions used in null score calculation.}

\item{ncores}{Number of cores for parallel processing.}
}
\value{
Modified `cbscore` matrix with transformed scores.
}
\description{
Transforms Cluster-Buster scores for enriched and depleted motifs.
}
