% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_peaks.R
\name{write_control_peak_to_bed}
\alias{write_control_peak_to_bed}
\title{Write Control Peaks to BED File}
\usage{
write_control_peak_to_bed(peak, fbed)
}
\arguments{
\item{peak}{A tibble containing peak data with columns: 'chr', 'start', and 'end'.}

\item{fbed}{A character string specifying the file path to write the BED file.}
}
\value{
None. The function writes the BED file to the specified path.
}
\description{
This function converts a tibble of control peaks into BED format and writes it to a specified file.
}
