% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_control_regions.R
\name{write_control_regions_to_bed}
\alias{write_control_regions_to_bed}
\title{Write control regions to BED File}
\usage{
write_control_regions_to_bed(control_region, fbed)
}
\arguments{
\item{control_region}{A tibble containing control region data with columns: 'chr', 'start', and 'end'.}

\item{fbed}{A character string specifying the file path to write the BED file.}
}
\value{
None. The function writes the BED file to the specified path.
}
\description{
This function converts a tibble of control regions into BED format and writes it to a specified file.
}
