% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrow_peaks.R
\name{write_narrow_peaks}
\alias{write_narrow_peaks}
\title{Write Narrow Peaks File}
\usage{
write_narrow_peaks(x, file)
}
\arguments{
\item{x}{A tibble containing the narrow peaks data.}

\item{file}{A string specifying the path to the output file.}
}
\value{
None
}
\description{
This function writes a tibble to a narrow peaks file.
}
