% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2gene_distance.R
\name{calc_peak2gene_distance}
\alias{calc_peak2gene_distance}
\title{Calculate Peak to Gene Distance}
\usage{
calc_peak2gene_distance(
  peak,
  ncores,
  meta_gene = grch38,
  max_distance = 5e+05,
  verbose = TRUE
)
}
\arguments{
\item{peak}{A tibble containing peak data with columns: 'chr', 'start', and 'end'.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{meta_gene}{A tibble containing gene metadata with columns: 'id', 'symbol', 'biotype', 'chr', 'start', 'end', 'strand', 'tss' (default is grch38).}

\item{max_distance}{A numeric value specifying the maximum distance to consider (default is 500,000).}

\item{verbose}{A logical value indicating whether to print progress messages (default is TRUE).}
}
\value{
A tibble with columns: 'gene', 'peak', and 'distance'.
}
\description{
This function calculates the distance between the center of peaks and the nearest gene transcription start site (TSS)
if the distance is less than a specified maximum distance.
}
