% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2gene.R
\name{calc_peak2gene_distance}
\alias{calc_peak2gene_distance}
\title{Calculate the distance between consensus peaks to genes}
\usage{
calc_peak2gene_distance(
  consensus_peak,
  ncores,
  meta_gene = TF2gene::grch38,
  max_distance = 5e+05
)
}
\arguments{
\item{consensus_peak}{A tibble containing consensus peak data with columns: 'chr', 'start', and 'end'.}

\item{ncores}{Number of cores to use for parallel processing.}

\item{meta_gene}{A tibble containing gene metadata with columns: 'id', 'symbol', 'biotype', 'chr', 'start', 'end', 'strand', 'tss' (default is TF2gene::grch38).}

\item{max_distance}{Maximum distance to consider (default is 500,000).}
}
\value{
A sparse matrix representing the distances. To distinguish distance 0 from empty sparse matrix
elements, distance 0 is set to 0.1.
}
\description{
This function calculates the distance between the center of peaks and the nearest TSS,
if the distance is less than a specified maximum distance.
}
