% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_active_motifs.R
\name{select_active_motifs}
\alias{select_active_motifs}
\title{Select Active Motifs Based on Transformed Cluster-Buster Scores}
\usage{
select_active_motifs(trans_cbscore, n_control)
}
\arguments{
\item{trans_cbscore}{Matrix of transformed Cluster-Buster scores (rows are motifs).}

\item{n_control}{Numeric, the number of control regions used.}
}
\value{
A tibble with motifs and their combined p-values.
}
\description{
This function applies the Cauchy Combination Test to transformed Cluster-Buster scores and
returns a tibble of motifs and combined p-values.
}
